refactor: use Sidebar menu on mobile
