+++ src/features/menu/MainMenu.tsx
@@ -1,59 +1,10 @@
-import React, { useState } from 'react'
+import React from 'react'
 import { Container, Icon, Image, Menu } from 'semantic-ui-react'
-import { Link, useHistory } from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import NetworkSelector from '../../components/NetworkSelector'
-import WalletSelector from '../onboard/walletSelector'
-import { RootState } from '../../app/rootReducer'
-import { useDispatch, useSelector } from 'react-redux'
-import WalletConfigModal from '../onboard/WalletConfigModal'
-import { selectWallet } from '../onboard/onboardSlice'
+import WalletSelectorContainer from '../onboard/WalletSelectorContainer'
 
 const MainMenu: React.FC = () => {
-    const [showWalletConfig, setShowWalletConfig] = useState(false)
-    const dispatch = useDispatch()
-    const history = useHistory()
-    const { wallet, onboardAPI, networkId } = useSelector(
-        (state: RootState) => state.onboard
-    )
-    const walletAddress = useSelector((state: RootState) => {
-        if (state.addresses.walletAddressId) {
-            return state.addresses.addressesById[
-                state.addresses.walletAddressId
-            ]
-        } else {
-            return undefined
-        }
-    })
-    const { mobile } = useSelector((state: RootState) => state.respsonsive)
-
-    const handleWalletConfig = () => {
-        setShowWalletConfig(true)
-    }
-
-    const handleCloseWalletConfig = () => {
-        setShowWalletConfig(false)
-    }
-
-    const handleSelectWallet = () => {
-        setShowWalletConfig(false)
-        dispatch(selectWallet(history))
-    }
-
-    const handleSelectAddress = () => {
-        setShowWalletConfig(false)
-        onboardAPI?.accountSelect()
-    }
-
-    const walletSelectorItem = wallet ? (
-        <Menu.Item>
-            <WalletSelector
-                handleClick={handleWalletConfig}
-                walletName={wallet?.name}
-                walletAccount={walletAddress}
-            />
-        </Menu.Item>
-    ) : null
-
     return (
         <Menu fixed='top' inverted size='huge'>
             <Container>
@@ -70,7 +21,9 @@ const MainMenu: React.FC = () => {
                     <Icon name='search' size={'large'} /> Check Allowances
                 </Menu.Item>
                 <Menu.Menu position='right'>
-                    {walletSelectorItem}
+                    <Menu.Item>
+                        <WalletSelectorContainer />
+                    </Menu.Item>
                     <Menu.Item>
                         <NetworkSelector />
                     </Menu.Item>
@@ -86,17 +39,6 @@ const MainMenu: React.FC = () => {
                     </Menu.Item>
                 </Menu.Menu>
             </Container>
-            {showWalletConfig && (
-                <WalletConfigModal
-                    handleClose={handleCloseWalletConfig}
-                    handleChangeWallet={handleSelectWallet}
-                    handleChangeAddress={handleSelectAddress}
-                    mobile={mobile}
-                    networkId={networkId}
-                    wallet={wallet}
-                    walletAddress={walletAddress}
-                />
-            )}
         </Menu>
     )
 }
